Revert "refactor: architecture DDD"
