+++ package.json
@@ -34,8 +34,6 @@
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
     "figlet": "^1.5.2",
-    "firebase-admin": "^11.10.1",
-    "googleapis": "^123.0.0",
     "js-yaml": "^4.1.0",
     "nest-commander": "^3.3.0",
     "ora": "^3.2.0",
