+++ src/app.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
-import { ApiModule } from './presentation/api/api.module';
+import { ApiModule } from './api/api.module';
 import { ApplicationModule } from './application/application.module';
 import { InfrastructureModule } from './infrastructure/infrastructure.module';
-import { ConfigurationModule } from './infrastructure/configuration/configuration.module';
-import { CliModule } from './presentation/cli/cli.module';
-import { GrpcModule } from './presentation/grpc/grpc.module';
+import { ConfigurationModule } from './configuration/configuration.module';
+import { CliModule } from './cli/cli.module';
+import { GrpcModule } from './grpc/grpc.module';
 
 @Module({
   imports: [
