+++ src/application/email/interfaces/email-service.interface.ts
