+++ src/application/email/models/email.model.ts
