+++ src/application/email/models/receiver.model.ts
