+++ src/cli/helpers/spinner.helper.ts
